fix schema

+++ db/migrate/20240203211758_create_merchants.rb
@@ -4,7 +4,7 @@ class CreateMerchants < ActiveRecord::Migration[7.0]
       t.integer :number_transactions, null: false, default: 0
       t.boolean :block_transactions, null: false, default: false
       t.integer :number_chargebacks, null: false, default: 0
-      t.float :total_value, null: false, default: 0.0
+      t.decimal :total_value, null: false, default: 0.0
 
       t.timestamps
     end
