+++ db/migrate/20240205193701_add_max_transaction_amount_to_user.rb
@@ -1,5 +1,5 @@
 class AddMaxTransactionAmountToUser < ActiveRecord::Migration[7.0]
   def change
-    add_column :users, :max_transaction_amount, :float
+    add_column :users, :max_transaction_amount, :decimal
   end
 end
